added auto close checkbox

+++ main.pas
@@ -1,4 +1,4 @@
-﻿unit main;
+unit main;
 
 interface
 
@@ -9,7 +9,7 @@ uses
   System.SysUtils, System.Variants, System.Classes,
   Vcl.Controls, Vcl.Forms, Vcl.Dialogs, Vcl.ExtCtrls, Vcl.Graphics,
   Vcl.StdCtrls, Vcl.Mask,
-  Winapi.Windows, Winapi.Messages;
+  Winapi.Windows, Winapi.Messages, RzRadChk;
 
 type
   TfrmMain = class(TForm)
@@ -24,10 +24,12 @@ type
     btnClose: TRzButton;
     btnCopy: TRzButton;
     btnOpen: TRzButton;
+    chkAutoClose: TRzCheckBox;
     procedure editTrackingNumberChange(Sender: TObject);
     procedure FormActivate(Sender: TObject);
     procedure btnCloseClick(Sender: TObject);
     procedure btnOpenClick(Sender: TObject);
+    procedure btnCopyClick(Sender: TObject);
   private
     { Private declarations }
   public
@@ -42,11 +44,25 @@ implementation
 
 {$R *.dfm}
 
+procedure TfrmMain.btnCopyClick(Sender: TObject);
+begin
+  // Copy the URL to the clipboard
+  Clipboard.AsText := trackingUrl;
+  if frmMain.chkAutoClose.Checked then begin
+    // Close the app because we're done here
+    Application.Terminate();
+  end;
+end;
+
 procedure TfrmMain.btnOpenClick(Sender: TObject);
 begin
   if Length(trackingUrl) > 0 then begin
+    // Open the default browser with the tracking URL
     ShellExecute(0, 'open', PChar(trackingUrl), '', '', SW_SHOWNORMAL);
-    Application.Terminate();
+    if frmMain.chkAutoClose.Checked then begin
+      // Close the app because we're done here
+      Application.Terminate();
+    end;
   end else begin
     showMessage('Nothing to do here, the tracking URL is empty');
   end;
@@ -59,16 +75,19 @@ end;
 
 procedure TfrmMain.editTrackingNumberChange(Sender: TObject);
 var
-  haveTrackingNumber: Boolean;
+  hasTrackingNumber: Boolean;
   trackingNumber: String;
 begin
   // Input field value changed, get the new tracking number
   trackingNumber := editTrackingNumber.Text;
-  haveTrackingNumber := Length(trackingNumber) > 0;
-  btnCopy.Visible := haveTrackingNumber;
-  btnOpen.Visible := haveTrackingNumber;
+  // Do we have a tracking number?
+  hasTrackingNumber := Length(trackingNumber) > 0;
+  // Updated the UI based on it existing
+  btnCopy.Visible := hasTrackingNumber;
+  btnOpen.Visible := hasTrackingNumber;
 
-  if haveTrackingNumber then begin
+  // Now update the tracking URL if we have a tracking number
+  if hasTrackingNumber then begin
     trackingUrl :=
       'https://www.ups.com/track?loc=en_US&requester=QUIC&tracknum=' +
       trackingNumber + '/trackdetails';
